move all_to_value to new utils.quantities module
