+++ ctapipe/image/muon/fitting.py
@@ -5,6 +5,8 @@ import scipy.constants as const
 from scipy.stats import norm
 from astropy.units import Quantity
 
+from ctapipe.utils.quantities import all_to_value
+
 __all__ = [
     'kundu_chaudhuri_circle_fit',
     'taubin_circle_fit',
@@ -86,24 +88,6 @@ def _psf_neg_log_likelihood(params, x, y, weights):
     return np.sum(
         (np.log(sigma) + 0.5 * ((pixel_distance - radius) / sigma)**2) * weights
     )
-
-def all_to_value(*args, unit):
-    '''converts all args to value if convertible to the same unit.
-
-    - does not copy the data
-    - the unit returned, will be the one of the 1st arg
-    - makes sure all args are convertible to the same unit
-    - return the values of all args and the unit
-    - Raise a meaningful error in case the args are not of a convertible unit.
-
-    Returns: *args_without_unit
-    '''
-    return tuple(
-        Quantity(arg, copy=False).to_value(unit)
-        for arg in args
-    )
-
-
 
 def psf_likelihood_fit(x, y, weights):
     """
